+++ examples/advanced/src/gen/zod/apiResponseSchema.ts
@@ -1,9 +1,11 @@
+import type { ApiResponse } from '../models/ts/ApiResponse.ts'
+import type { ToZod } from '@kubb/plugin-zod/utils'
 import { z } from 'zod'
 
 export const apiResponseSchema = z.object({
   code: z.number().int().optional(),
   type: z.string().optional(),
   message: z.string().optional(),
-})
+} satisfies ToZod<ApiResponse>)
 
-export type ApiResponseSchema = z.infer<typeof apiResponseSchema>
+export type ApiResponseSchema = ApiResponse