+++ examples/advanced/src/gen/zod/categorySchema.ts
@@ -1,8 +1,10 @@
+import type { Category } from '../models/ts/Category.ts'
+import type { ToZod } from '@kubb/plugin-zod/utils'
 import { z } from 'zod'
 
 export const categorySchema = z.object({
   id: z.number().int().optional(),
   name: z.string().optional(),
-})
+} satisfies ToZod<Category>)
 
-export type CategorySchema = z.infer<typeof categorySchema>
+export type CategorySchema = Category