+++ examples/advanced/src/gen/zod/petNotFoundSchema.ts
@@ -1,8 +1,10 @@
+import type { PetNotFound } from '../models/ts/PetNotFound.ts'
+import type { ToZod } from '@kubb/plugin-zod/utils'
 import { z } from 'zod'
 
 export const petNotFoundSchema = z.object({
   code: z.number().int().optional(),
   message: z.string().optional(),
-})
+} satisfies ToZod<PetNotFound>)
 
-export type PetNotFoundSchema = z.infer<typeof petNotFoundSchema>
+export type PetNotFoundSchema = PetNotFound