+++ examples/advanced/src/gen/zod/tag/tagSchema.ts
@@ -1,8 +1,10 @@
+import type { TagTag } from '../../models/ts/tag/Tag.ts'
+import type { ToZod } from '@kubb/plugin-zod/utils'
 import { z } from 'zod'
 
 export const tagTagSchema = z.object({
   id: z.number().int().optional(),
   name: z.string().optional(),
-})
+} satisfies ToZod<TagTag>)
 
-export type TagTagSchema = z.infer<typeof tagTagSchema>
+export type TagTagSchema = TagTag