+++ examples/advanced/src/gen/zod/userController/createUserSchema.ts
@@ -1,3 +1,4 @@
+import type { CreateUserError, CreateUserMutationRequest, CreateUserMutationResponse } from '../../models/ts/userController/CreateUser.ts'
 import { userSchema } from '../userSchema.ts'
 import { z } from 'zod'
 
@@ -6,15 +7,15 @@ import { z } from 'zod'
  */
 export const createUserErrorSchema = z.lazy(() => userSchema)
 
-export type CreateUserErrorSchema = z.infer<typeof createUserErrorSchema>
+export type CreateUserErrorSchema = CreateUserError
 
 /**
  * @description Created user object
  */
 export const createUserMutationRequestSchema = z.lazy(() => userSchema)
 
-export type CreateUserMutationRequestSchema = z.infer<typeof createUserMutationRequestSchema>
+export type CreateUserMutationRequestSchema = CreateUserMutationRequest
 
 export const createUserMutationResponseSchema = z.any()
 
-export type CreateUserMutationResponseSchema = z.infer<typeof createUserMutationResponseSchema>
+export type CreateUserMutationResponseSchema = CreateUserMutationResponse