+++ examples/advanced/src/gen/zod/userController/createUsersWithListInputSchema.ts
@@ -1,3 +1,9 @@
+import type {
+  CreateUsersWithListInput200,
+  CreateUsersWithListInputError,
+  CreateUsersWithListInputMutationRequest,
+  CreateUsersWithListInputMutationResponse,
+} from '../../models/ts/userController/CreateUsersWithListInput.ts'
 import { userSchema } from '../userSchema.ts'
 import { z } from 'zod'
 
@@ -6,19 +12,19 @@ import { z } from 'zod'
  */
 export const createUsersWithListInput200Schema = z.lazy(() => userSchema)
 
-export type CreateUsersWithListInput200Schema = z.infer<typeof createUsersWithListInput200Schema>
+export type CreateUsersWithListInput200Schema = CreateUsersWithListInput200
 
 /**
  * @description successful operation
  */
 export const createUsersWithListInputErrorSchema = z.any()
 
-export type CreateUsersWithListInputErrorSchema = z.infer<typeof createUsersWithListInputErrorSchema>
+export type CreateUsersWithListInputErrorSchema = CreateUsersWithListInputError
 
 export const createUsersWithListInputMutationRequestSchema = z.array(z.lazy(() => userSchema))
 
-export type CreateUsersWithListInputMutationRequestSchema = z.infer<typeof createUsersWithListInputMutationRequestSchema>
+export type CreateUsersWithListInputMutationRequestSchema = CreateUsersWithListInputMutationRequest
 
 export const createUsersWithListInputMutationResponseSchema = z.lazy(() => createUsersWithListInput200Schema)
 
-export type CreateUsersWithListInputMutationResponseSchema = z.infer<typeof createUsersWithListInputMutationResponseSchema>
+export type CreateUsersWithListInputMutationResponseSchema = CreateUsersWithListInputMutationResponse