+++ packages/plugin-zod/src/generators/operationsGenerator.test.tsx
@@ -3,7 +3,7 @@ import { createMockedPluginManager, matchFiles, mockedPluginManager } from '@kub
 import path from 'node:path'
 import type { Plugin } from '@kubb/core'
 import type { HttpMethod } from '@kubb/oas'
-import { parse } from '@kubb/oas/parser'
+import { parse } from '@kubb/oas'
 import { OperationGenerator } from '@kubb/plugin-oas'
 import type { PluginZod } from '../types.ts'
 import { operationsGenerator } from './operationsGenerator.tsx'