+++ packages/plugin-zod/src/utils/index.ts
@@ -1,0 +1,1 @@
+export type { ToZod } from './ToZod.ts'